+++ test/mitmproxy/addons/test_proxyauth.py
@@ -7,16 +7,9 @@ from mitmproxy.test import tutils
 from mitmproxy.addons import proxyauth
 
 
-def mkauth(username, password, scheme="basic"):
-    v = binascii.b2a_base64(
-        (username + ":" + password).encode("utf8")
-    ).decode("ascii")
-    return scheme + " " + v
-
-
 def test_parse_http_basic_auth():
     assert proxyauth.parse_http_basic_auth(
-        mkauth("test", "test")
+        proxyauth.mkauth("test", "test")
     ) == ("basic", "test", "test")
     assert not proxyauth.parse_http_basic_auth("")
     assert not proxyauth.parse_http_basic_auth("foo bar")
@@ -92,19 +85,23 @@ def test_check():
         ctx.configure(up, auth_nonanonymous=True)
         f = tflow.tflow()
         assert not up.check(f)
-        f.request.headers["Proxy-Authorization"] = mkauth("test", "test")
+        f.request.headers["Proxy-Authorization"] = proxyauth.mkauth(
+            "test", "test"
+        )
         assert up.check(f)
 
         f.request.headers["Proxy-Authorization"] = "invalid"
         assert not up.check(f)
 
-        f.request.headers["Proxy-Authorization"] = mkauth(
+        f.request.headers["Proxy-Authorization"] = proxyauth.mkauth(
             "test", "test", scheme = "unknown"
         )
         assert not up.check(f)
 
         ctx.configure(up, auth_nonanonymous=False, auth_singleuser="test:test")
-        f.request.headers["Proxy-Authorization"] = mkauth("test", "test")
+        f.request.headers["Proxy-Authorization"] = proxyauth.mkauth(
+            "test", "test"
+        )
         assert up.check(f)
         ctx.configure(up, auth_nonanonymous=False, auth_singleuser="test:foo")
         assert not up.check(f)
@@ -116,9 +113,13 @@ def test_check():
                 "mitmproxy/net/data/htpasswd"
             )
         )
-        f.request.headers["Proxy-Authorization"] = mkauth("test", "test")
+        f.request.headers["Proxy-Authorization"] = proxyauth.mkauth(
+            "test", "test"
+        )
         assert up.check(f)
-        f.request.headers["Proxy-Authorization"] = mkauth("test", "foo")
+        f.request.headers["Proxy-Authorization"] = proxyauth.mkauth(
+            "test", "foo"
+        )
         assert not up.check(f)
 
 
@@ -133,7 +134,9 @@ def test_authenticate():
         assert f.response.status_code == 407
 
         f = tflow.tflow()
-        f.request.headers["Proxy-Authorization"] = mkauth("test", "test")
+        f.request.headers["Proxy-Authorization"] = proxyauth.mkauth(
+            "test", "test"
+        )
         up.authenticate(f)
         assert not f.response
         assert not f.request.headers.get("Proxy-Authorization")
@@ -146,7 +149,9 @@ def test_authenticate():
 
         f = tflow.tflow()
         f.mode = "transparent"
-        f.request.headers["Authorization"] = mkauth("test", "test")
+        f.request.headers["Authorization"] = proxyauth.mkauth(
+            "test", "test"
+        )
         up.authenticate(f)
         assert not f.response
         assert not f.request.headers.get("Authorization")
